handle lunr fields in index config, drop WaxTasks::LunrCollection
