+++ lib/wax_tasks/lunr_collection.rb
@@ -30,41 +30,6 @@ module WaxTasks
       raise Error::WaxTasksError, "Label file #{label_path} could not be loaded. Make sure it is a valid YAML file.\n#{e}"
     end
 
-    # Finds the page_dir of markdown pages for the collection and ingests
-    # them as an array of hashes
-    #
-    # @return [Array] array of the loaded markdown pages loaded as hashes
-    def ingest_pages
-      data  = []
-      pages = Dir.glob("#{self.page_dir}/*.md")
-      puts "There are no pages in #{self.page_dir} to index.".cyan if pages.empty?
-      pages.each do |p|
-        begin
-          data << load_page(p)
-        rescue StandardError => e
-          raise Error::LunrPageLoad, "Cannot load page #{p}\n#{e}"
-        end
-      end
-      data
-    end
-
-    # Reads in a markdown file and converts it to a hash
-    # with the values from @fields.
-    # Adds the content of the file (below the YAML) if @content == true
-    #
-    # @param page [String] the path to a markdown page to load
-    def load_page(page)
-      yaml = YAML.load_file(page)
-      hash = {
-        'link' => "{{'#{yaml.fetch('permalink')}' | relative_url }}",
-        'collection' => @name
-      }
-      content = WaxTasks::Utils.html_strip(File.read(page))
-      hash['content'] = WaxTasks::Utils.remove_diacritics(content)
-      fields = @fields.push('pid').uniq
-      fields.push('thumbnail') if yaml.key?('thumbnail')
-      fields.each { |f| hash[f] = yaml[f].lunr_normalize }
-      hash
-    end
+   
   end
 end
